Refactoring things.
